Rename Gameboard, move createPlayer, Create Cell
Move createPlayer into the newly renamed Gameboard as players will only
be needed inside the Gameboard instance.
Create a for loop that fills the gamesBoard with appropriate cells,
which is just now declared but is missing its functional code, these
Cell's are important as they will reside in arrays as their elements
and so playing to a cell calls the appropriate namespace for that tile
on the gameBoard.
I need to rename gameBoard to just board.