Add functionality for PlayGame variables, function
For our PlayGame to work correctly we need to create placeholders for certain values like how many moves have been made so we can stop the game when the board is full, nextPlayerMove so we can keep track of who's turn it is next, players object so we have specific values saved and easily accessible for when a player sets themselves in a Cell.
I have just started the function to ask a player for a move but need to go back and change the array to remain just an array of elements and not an array of arrays of elements as it will make checking for winning combinations way easier.