Change board value to 9 Cell element array
This should help us later with finding winning combinations easier than what it would have been, and changed the function to return an object with this board, so that we can now access this inside our PlayGame function, for which we renamed inside PlayGame for more clarity.
Declared a function that will output into the console our array in a more user friendly manner, but this function is missing its code for now, that'll be added next.