Change tie condition, and while conditions to play
The conditions for a tie were wrong in that it took more moves than are actually played to stop the gameplay loop, as well as it had required the max moves to be had and a winner, now it requires either 9 moves to have been taken or a winner.
Likewise the conditions for a tie to be announced were wrong as well, it used a condition of greater than 9 moves which is wrong because that means it wouldn't announce a tie until a move after the board was full which is impossible.