Add function for setting markers, fix condition
By moving this chunk of code into a container, we not only are organizing our code better but have created a reusable action which could be beneficial elsewhere.
Added a comment to explain what the while loop is doing.
There was still a problem with tie games, but this time it's because I changed the operator to or, which is wrong because then it would look for if either of the two conditions were true and loop, I want it to stop running if either are true so I changed it to an and.