Change checkWinner to accept an array
By changing this function from calling on the board property of the PlayGame functions interface we allow the function to work on any array of 9 elements representing a tictactoe board. Of course these boards will have to follow the setup we have, _ are the starting markers and X and O are the player markers.