Add styling to h1, .marker
The h1 needs a larger font size than other text on the screen, of course not bigger than the text which is representative of the markers on the board.
I also changed the .marker elements to look more like gameboard marker locations by changing their font size and background colors, as well as changing them to inline display which div's are not by default so I could lay the elements out more appropriately.
Also added a gap to the container grid so its elements wouldn't be so snug up to each other.