Change askPlayerMove to a console.log
Instead of using prompt we'll have this function print off the current tictactoe board in console and log a statement that it's a specific players turn to place their marker, aka use the setPlayerMarker function themselves.
Also changed the variable name we were passing inside the setPlayerMarker function as it was using an old value and somehow was still working.
I also started working on the playTicTacToe function but ran into a bunch of issues as we haven't learned promises or async await yet in the course so I have to think about how I want to attempt this, I want to call askPlayerMove and then have a listener persay wait until they make a valid move.