Remove playTicTacToe, change setPlayerMarker
I got ahead of myself in the directions, and should have been focusing mvp at each stage of the projects directions.
I have commented out playTicTacToe function and instead am returning setPlayerMarker to the obj returned by the PlayGame module pattern which allows players to play the game in the console using the methods of the module.