Change markPositions to use argument array
With this change the function now just uses the board from the GameBoard.board object instead of creating an unnecessary array.
Removed argument to displayGame module pattern function, it simply wasn't need nor was it being used.
Added this function being invocated when a player successfully sets their marker so that the board in the dom is repainted with the move.