Add variable and event listener for gameBoard
This listener watches for clicks on child elements in the #gameBoard element and bubbles to the appropriate .marker element that was clicked so we can mark that element with the players marker.
Also moved markerElements back to displayGame as it wasn't needed in PlayGame because of the above code, and with that I had to change arguments and exports again.
Also removed some white spacing that was used for cleanliness during pseudo code but is no longer needed, and removed pseudo code that is no longer needed as the code has been added.