Move displayPlayerNames invocation location
Before this function was being invoked inside the changePlayerNames function so for more clarity I have moved it into the event listeners block.
There is a possibility to have the change name function be workable in the console only version and call that function in the dom display but I won't do it because I'm already going beyond.