Remove unnecessary arg, use moves attr to stop
Found an argument to a getValue invocation which was doing nothing.
Changed interaction of setPlayerMarker for when board has a winning combination to not allow players to mark, as that was interfering with highlighting the correct winning combination.
However there are still problems with highlighting the correct combination of elements, mainly if there was a previous game and the restart button was pressed the winning combination of the new gam is not highlighted but the old game combination is.
Also omitted the if statement from the highlightWinningElements as this function is now strategically invoked, but I will look into it next.
Also exported the winningArr so the restart button can reset that array when playing a new game, although it's not working as of right now.